pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "courierdb"
version = "0.2.5"
description = "JSON-first database with API and MCP interfaces"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
keywords = ["database", "json", "api", "mcp", "agent"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi==0.124.0",
    "uvicorn[standard]==0.41.0",
    "pydantic==2.12.5",
    "lmdb==1.7.5",
    "requests==2.32.5",
    "fastmcp==2.13.3",
    "python-dotenv==1.2.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/courierdb"
Issues = "https://github.com/yourusername/courierdb/issues"

[project.scripts]
courierdb = "courierdb.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["courierdb"]
exclude = ["Dockerfile", "docker-compose.yml", ".env", "tests"]
